Move AccountSwitcher Logic Into Service
